requirements.txt
fastapi
uvicorn
websockets
python-dotenv
google-generativeai
deepgram-sdk
elevenlabs
twilio
psycopg2-binary